Rename the artifacts of multigroup using the group name
Single files are created by using group name and suffix, multiple files
are prefixed by group name

+++ de.dieklaut.camtool/src/de/dieklaut/camtool/operations/Render.java
@@ -8,6 +8,7 @@ import java.util.Calendar;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.Iterator;
 import java.util.Map;
 import java.util.Properties;
 import java.util.Set;
@@ -16,6 +17,7 @@ import de.dieklaut.camtool.Constants;
 import de.dieklaut.camtool.Context;
 import de.dieklaut.camtool.Group;
 import de.dieklaut.camtool.Logger;
+import de.dieklaut.camtool.MultiGroup;
 import de.dieklaut.camtool.Logger.Level;
 import de.dieklaut.camtool.Sorter;
 import de.dieklaut.camtool.SortingHelper;
@@ -77,6 +79,7 @@ public class Render extends AbstractOperation {
 
 		Map<RenderJob, String> renderJobToGroupName = new HashMap<>();
 		Map<RenderJob, String> renderJobToChecksum = new HashMap<>();
+		Map<RenderJob, String> renderJobToNamePrefix = new HashMap<>();
 
 		Set<Path> rendered = new HashSet<>();
 		rendered.add(sourceStatePath);
@@ -88,36 +91,103 @@ public class Render extends AbstractOperation {
 			String newChecksum = hasChanges(group, sourceState);
 			RenderJob renderJob = group.getRenderJob();
 			try {
-				Collection<? extends Path> predictedResults = renderJob.getPredictedResults(destination_direct);
+				Collection<Path> predictedResults = renderJob.getPredictedResults(destination_direct);
 				predictionFailed |= predictedResults == null;
 
-				if (force || newChecksum != null || predictionFailed) {
-					if (force) {
+				
+				Collection<Path> modifiedPaths = new HashSet<>();
+				if (group instanceof MultiGroup) {
+					if (predictedResults != null && predictedResults.size() == 1) {
+						Path result = predictedResults.iterator().next();
+						Path newPath = destination_direct.resolve(group.getName() + FileUtils.getSuffix(result));
+						modifiedPaths.add(newPath);
+						Logger.log("Modified predicted path of single Multigroup output " + result + " to " + newPath, Level.DEBUG);
+					} else if (predictedResults != null && predictedResults.size() > 1) {
+						Iterator<Path> it = predictedResults.iterator();
+						while (it.hasNext()) {
+							Path result = it.next();
+							Path newPath = destination_direct.resolve(group.getName() + "_" + result.getFileName().toString());
+							modifiedPaths.add(newPath);
+							Logger.log("Modified predicted path of single Multigroup output " + result + " to " + newPath, Level.DEBUG);
+						}
+					}
+					predictedResults = modifiedPaths;
+				}
+				
+				boolean predictedFilesNotExisting = false;
+				for (Path p : predictedResults) {
+					if (!Files.exists(p)) {
+						predictedFilesNotExisting = true;
+						break;
+					}
+				}
+				
+				boolean executeForced = force || predictionFailed || predictedFilesNotExisting;
+				
+				if (executeForced || newChecksum != null) {
+					if (executeForced) {
 						newChecksum = FileUtils.getChecksum(group.getAllFiles());
 					}
 					renderJobToGroupName.put(renderJob, group.getName());
 					renderJobToChecksum.put(renderJob, newChecksum);
+					if (group instanceof MultiGroup) {
+						renderJobToNamePrefix.put(renderJob, group.getName());
+					}
 				} else {
 					rendered.addAll(predictedResults);
 				}
+				
+
+				
 			} catch (IOException e) {
 				Logger.log("Render job creation failed", e);
 				errorDuringJobCreations = true;
 			}
 		}
 
+		Logger.log("Rendering " + renderJobToGroupName.size() + " jobs.", Level.INFO);
+		
 		for (RenderJob job : renderJobToGroupName.keySet()) {
 			try {
 				sourceState.setProperty(renderJobToGroupName.get(job), renderJobToChecksum.get(job));
-				rendered.addAll(job.store(destination_direct));
+				Set<Path> jobResult = job.store(destination_direct);
+				
+				if (renderJobToNamePrefix.containsKey(job) && jobResult.size() == 1) {
+					Path result = jobResult.iterator().next();
+					Path newPath = destination_direct.resolve(renderJobToNamePrefix.get(job) + FileUtils.getSuffix(result));
+					Files.deleteIfExists(newPath);
+					Files.move(result, newPath);
+					jobResult.remove(result);
+					jobResult.add(newPath);
+					Logger.log("Renamed single Multigroup output " + result + " to " + newPath, Level.INFO);
+				} else if (renderJobToNamePrefix.containsKey(job) && jobResult.size() > 1) {
+					Set<Path> toBeRemoved = new HashSet<>();
+					Set<Path> toBeAdded = new HashSet<>();
+					Iterator<Path> it = jobResult.iterator();
+					while (it.hasNext()) {
+						Path result = it.next();
+						Path newPath = destination_direct.resolve(renderJobToNamePrefix.get(job) + "_" + result.getFileName().toString());
+						Files.deleteIfExists(newPath);
+						Files.move(result, newPath);
+						toBeRemoved.add(result);
+						toBeAdded.add(newPath);
+						Logger.log("Prefixed single Multigroup output " + result + " to " + newPath, Level.INFO);
+					}
+					jobResult.removeAll(toBeRemoved);
+					jobResult.addAll(toBeAdded);
+				}
+				
+				rendered.addAll(jobResult);
 				sourceState.store(Files.newOutputStream(sourceStatePath),
 						"Last update on " + Calendar.getInstance().getTime());
 			} catch (IOException e) {
 				Logger.log("Failed to execute render job " + job, e, Level.WARNING);
 			}
 		}
+		
+		Logger.log("Rendered finished: " + (rendered.size() - 1) + " resulting artifacts.", Level.INFO); //-1 because of .sourcestate file
 
-		if (!errorDuringJobCreations) {
+		if (!errorDuringJobCreations && groupName == null) {
 			try {
 				FileUtils.deleteEverythingBut(destination_direct, rendered);
 			} catch (IOException e) {
